Installation script for Mac users
New installation script to be provided to users with Mac local wallets as they usually install the latest version or their VPS 
and then have to downgrade the VPS wallet in order to get MN working. 
The only change done is in wallet version to be downloaded by the script.